Conflicts:
	build.gradle
	build.properties
	src/main/java/mhfc/net/MHFCMain.java
	src/main/java/mhfc/net/client/model/weapon/bow/ModelBPower.java
	src/main/java/mhfc/net/common/core/registry/MHFCEventRegistry.java
	src/main/java/mhfc/net/common/eventhandler/player/ConnectionEventHandler.java
	src/main/java/mhfc/net/common/system/UpdateSystem.java
	src/main/java/mhfc/net/common/util/SubTypedItem.java
	src/main/java/mhfc/net/common/util/Utilities.java
	src/main/java/mhfc/net/common/util/gui/MHFCGuiUtil.java
	src/main/java/mhfc/net/common/util/lib/MHFCReference.java
	src/main/java/mhfc/net/common/weapon/range/iBowAnim.java
	src/main/resources/mcmod.info

+++ src/main/java/mhfc/net/common/util/gui/MHFCGuiUtil.java
@@ -18,15 +18,13 @@ public class MHFCGuiUtil {
 
 	public static int realScreenWidth(Minecraft mc) {
 		if (mc == null)
-			throw new IllegalArgumentException(
-				"Gui utils may only be accessed with valid minecraft");
+			throw new IllegalArgumentException("Gui utils may only be accessed with valid minecraft");
 		return mc.displayWidth;
 	}
 
 	public static int realScreenHeight(Minecraft mc) {
 		if (mc == null)
-			throw new IllegalArgumentException(
-				"Gui utils may only be accessed with valid minecraft");
+			throw new IllegalArgumentException("Gui utils may only be accessed with valid minecraft");
 		return mc.displayHeight;
 	}
 
@@ -47,10 +45,8 @@ public class MHFCGuiUtil {
 
 	private static void refreshScaled(Minecraft mc) {
 		if (mc == null)
-			throw new IllegalArgumentException(
-				"Gui utils may only be accessed with valid minecraft");
-		s = new ScaledResolution(mc, mc.displayWidth,
-			mc.displayHeight);
+			throw new IllegalArgumentException("Gui utils may only be accessed with valid minecraft");
+		s = new ScaledResolution(mc, mc.displayWidth, mc.displayHeight);
 	}
 
 	/**
@@ -63,12 +59,10 @@ public class MHFCGuiUtil {
 	 * @return The drawn height of the string. Always line height for valid
 	 *         parameters and width==0
 	 */
-	public static int drawTextAndReturnHeight(FontRenderer fRend,
-		String string, int posX, int posY, int width, int colour) {
+	public static int drawTextAndReturnHeight(FontRenderer fRend, String string, int posX, int posY, int width,
+			int colour) {
 		if (fRend == null || string == null) {
-			MHFCMain.logger.warn(fRend == null
-				? "Null renderer used as argument"
-				: "Render request for a null string");
+			MHFCMain.logger.warn(fRend == null ? "Null renderer used as argument" : "Render request for a null string");
 			Thread.dumpStack();
 			return 0;
 		}
@@ -94,72 +88,58 @@ public class MHFCGuiUtil {
 	 * @return If the font renderer will crash the game with this string and
 	 *         this width
 	 */
-	public static boolean isDrawWidthTooSmall(FontRenderer fRend, int width,
-		String string) {
-		return !string.isEmpty()
-			&& width < fRend.getStringWidth(string.substring(0, 1));
+	public static boolean isDrawWidthTooSmall(FontRenderer fRend, int width, String string) {
+		return !string.isEmpty() && width < fRend.getStringWidth(string.substring(0, 1));
 	}
 
-	public static void drawTexturedRectangle(double xMin, double yMin,
-		double width, double height, float u, float v, float uWidth,
-		float vHeight) {
+	public static void drawTexturedRectangle(double xMin, double yMin, double width, double height, float u, float v,
+			float uWidth, float vHeight) {
 		Tessellator t = Tessellator.instance;
 		t.startDrawingQuads();
 		t.addVertexWithUV(xMin, yMin, zLevel, u, v);
 		t.addVertexWithUV(xMin, yMin + height, zLevel, u, v + vHeight);
-		t.addVertexWithUV(xMin + width, yMin + height, zLevel, u + uWidth, v
-			+ vHeight);
+		t.addVertexWithUV(xMin + width, yMin + height, zLevel, u + uWidth, v + vHeight);
 		t.addVertexWithUV(xMin + width, yMin, zLevel, u + uWidth, v);
 		t.draw();
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, int width,
-		int height) {
+	public static void drawTexturedBoxFromBorder(float x, float y, float width, float height) {
 		drawTexturedBoxFromBorder(x, y, zLevel, width, height);
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, float zLevel,
-		int width, int height) {
-		drawTexturedBoxFromBorder(x, y, zLevel, width, height, Math.min(Math
-			.min(15, width / 2), height / 2));
+	public static void drawTexturedBoxFromBorder(float x, float y, float zLevel, float width, float height) {
+		drawTexturedBoxFromBorder(x, y, zLevel, width, height, Math.min(Math.min(15, width / 2), height / 2));
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, float zLevel,
-		int width, int height, int borderSize) {
-		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize,
-			borderSize / 256f);
+	public static void drawTexturedBoxFromBorder(float x, float y, float zLevel, float width, float height,
+			float borderSize) {
+		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize, borderSize / 256f);
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, float zLevel,
-		int width, int height, int borderSize, float borderUV) {
-		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize,
-			borderUV, 1, 1);
+	public static void drawTexturedBoxFromBorder(float x, float y, float zLevel, float width, float height,
+			float borderSize, float borderUV) {
+		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize, borderUV, 1, 1);
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, float zLevel,
-		int width, int height, int borderSize, float borderUV, float maxU,
-		float maxV) {
-		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize,
-			borderUV, borderUV, maxU, maxV);
+	public static void drawTexturedBoxFromBorder(float x, float y, float zLevel, float width, float height,
+			float borderSize, float borderUV, float maxU, float maxV) {
+		drawTexturedBoxFromBorder(x, y, zLevel, width, height, borderSize, borderUV, borderUV, maxU, maxV);
 	}
 
-	public static void drawTexturedBoxFromBorder(int x, int y, float zLevel,
-		int width, int height, int borderSize, float borderU, float borderV,
-		float maxU, float maxV) {
+	public static void drawTexturedBoxFromBorder(float x, float y, float zLevel, float width, float height,
+			float borderSize, float borderU, float borderV, float maxU, float maxV) {
 		Tessellator tess = Tessellator.instance;
 		tess.startDrawingQuads();
 		tess.addVertexWithUV(x, y, zLevel, 0, 0);
 		tess.addVertexWithUV(x, y + borderSize, zLevel, 0, borderV);
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU,
-			borderV);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU, borderV);
 		tess.addVertexWithUV(x + borderSize, y, zLevel, borderU, 0);
 		tess.draw();
 		tess.startDrawingQuads();
 		tess.addTranslation(width - borderSize, 0, 0);
 		tess.addVertexWithUV(x, y, zLevel, maxU - borderU, 0);
 		tess.addVertexWithUV(x, y + borderSize, zLevel, maxU - borderU, borderV);
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, maxU,
-			borderV);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, maxU, borderV);
 		tess.addVertexWithUV(x + borderSize, y, zLevel, maxU, 0);
 		tess.draw();
 		tess.startDrawingQuads();
@@ -173,61 +153,44 @@ public class MHFCGuiUtil {
 		tess.addTranslation(-width + borderSize, 0, 0);
 		tess.addVertexWithUV(x, y, zLevel, 0, maxV - borderV);
 		tess.addVertexWithUV(x, y + borderSize, zLevel, 0, maxV);
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU,
-			maxV);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU, maxV);
 		tess.addVertexWithUV(x + borderSize, y, zLevel, borderU, maxV - borderV);
 		tess.draw();
 		tess.addTranslation(0, -height + borderSize, 0);
 
 		tess.startDrawingQuads();
 		tess.addVertexWithUV(x, y + borderSize, zLevel, 0, borderV);
-		tess.addVertexWithUV(x, y + height - borderSize, zLevel, 0, maxV
-			- borderV);
-		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel,
-			borderU, maxV - borderV);
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU,
-			borderV);
+		tess.addVertexWithUV(x, y + height - borderSize, zLevel, 0, maxV - borderV);
+		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel, borderU, maxV - borderV);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU, borderV);
 		tess.draw();
 
 		tess.startDrawingQuads();
-		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel,
-			maxU - borderU, borderV);
-		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize,
-			zLevel, maxU - borderU, maxV - borderV);
-		tess.addVertexWithUV(x + width, y + height - borderSize, zLevel, maxU,
-			maxV - borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel, maxU - borderU, borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize, zLevel, maxU - borderU, maxV - borderV);
+		tess.addVertexWithUV(x + width, y + height - borderSize, zLevel, maxU, maxV - borderV);
 		tess.addVertexWithUV(x + width, y + borderSize, zLevel, maxU, borderV);
 		tess.draw();
 
 		tess.startDrawingQuads();
 		tess.addVertexWithUV(x + borderSize, y, zLevel, borderU, 0);
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU,
-			borderV);
-		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel,
-			maxU - borderU, borderV);
-		tess.addVertexWithUV(x + width - borderSize, y, zLevel, maxU - borderU,
-			0);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU, borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel, maxU - borderU, borderV);
+		tess.addVertexWithUV(x + width - borderSize, y, zLevel, maxU - borderU, 0);
 		tess.draw();
 
 		tess.startDrawingQuads();
-		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel,
-			borderU, maxV - borderV);
+		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel, borderU, maxV - borderV);
 		tess.addVertexWithUV(x + borderSize, y + height, zLevel, borderU, maxV);
-		tess.addVertexWithUV(x + width - borderSize, y + height, zLevel, maxU
-			- borderU, maxV);
-		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize,
-			zLevel, maxU - borderU, maxV - borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + height, zLevel, maxU - borderU, maxV);
+		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize, zLevel, maxU - borderU, maxV - borderV);
 		tess.draw();
 
 		tess.startDrawingQuads();
-		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU,
-			borderV);
-		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel,
-			borderU, maxV - borderV);
-		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize,
-			zLevel, maxU - borderU, maxV - borderV);
-		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel,
-			maxU - borderU, borderV);
+		tess.addVertexWithUV(x + borderSize, y + borderSize, zLevel, borderU, borderV);
+		tess.addVertexWithUV(x + borderSize, y + height - borderSize, zLevel, borderU, maxV - borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + height - borderSize, zLevel, maxU - borderU, maxV - borderV);
+		tess.addVertexWithUV(x + width - borderSize, y + borderSize, zLevel, maxU - borderU, borderV);
 		tess.draw();
 
 	}
